Refactor move files
