--- fixture-monkey/src/main/java/com/navercorp/fixturemonkey/builder/ArbitraryBuilderVisitor.java
@@ -1,5 +1,0 @@
-package com.navercorp.fixturemonkey.builder;
-
-public interface ArbitraryBuilderVisitor {
-	void visit(ArbitraryBuilder arbitraryBuilder);
-}
