+++ fixture-monkey/src/main/java/com/navercorp/fixturemonkey/customizer/ArbitraryBuilderVisitor.java
@@ -1,0 +1,7 @@
+package com.navercorp.fixturemonkey.customizer;
+
+import com.navercorp.fixturemonkey.builder.ArbitraryBuilder;
+
+public interface ArbitraryBuilderVisitor {
+	void visit(ArbitraryBuilder arbitraryBuilder);
+}
